pyproject.toml
[tool.poetry]
name = "cascabel"
version = "0.1.0"
description = ""
authors = ["Elijah J. Passmore <elijahjpassmore@elijahjpassmore.com>"]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"
click = "^8.0.4"
loguru = "^0.6.0"
types-PyYAML = "^6.0.4"
GitPython = "^3.1.27"

[tool.poetry.dev-dependencies]
mypy = "^0.931"
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
